Add built-in initial match loading
